+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/Payload.java
@@ -91,6 +91,7 @@ public abstract class Payload {
     public enum PayloadType {
         TRANSFER,
         TRANSFER_WITH_MEMO,
+        REGISTER_DATA,
         INITIALIZE_SMART_CONTRACT_INSTANCE,
     }