+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Unreleased changes
 
 - Fixed bug in `UpdateEnqueuedResult` which parsed `effectiveTime` in wrong.
+- Added support for `Uptime` query.
+- Added support for Account Transaction Register Data.
+- Added support for query `PeerTotalSent`
+- Added support for query `PeerList`
+- Added support for query `PeerStats`
+- Added support for query `PeerVersion`
+- Added support for query `NodeInfo`.
 
 ## 3.0.0
 - Stronger typing for various places in the API.